Moved chunk and voxel to core crate
