+++ crates/core/src/lib.rs
@@ -1,8 +1,5 @@
 pub mod chunk;
 pub mod landscape;
 pub mod math;
-pub mod query;
+// pub mod query;
 pub mod voxel;
-mod voxworld;
-
-pub use voxworld::VoxWorld;
