--- crates/world_server/src/chunk.rs
@@ -1,504 +1,0 @@
-use bevy::math::{IVec2, IVec3, Vec3};
-use projekto_core::{math, voxel};
-use serde::{Deserialize, Serialize};
-
-use crate::{Chunk, Voxel};
-
-pub const X_AXIS_SIZE: usize = 16;
-pub const Y_AXIS_SIZE: usize = 256;
-pub const Z_AXIS_SIZE: usize = 16;
-
-pub const X_END: i32 = (X_AXIS_SIZE - 1) as i32;
-pub const Y_END: i32 = (Y_AXIS_SIZE - 1) as i32;
-pub const Z_END: i32 = (Z_AXIS_SIZE - 1) as i32;
-
-pub const BUFFER_SIZE: usize = X_AXIS_SIZE * Z_AXIS_SIZE * Y_AXIS_SIZE;
-
-const X_SHIFT: usize = (Z_AXIS_SIZE.ilog2() + Z_SHIFT as u32) as usize;
-const Z_SHIFT: usize = Y_AXIS_SIZE.ilog2() as usize;
-const Y_SHIFT: usize = 0;
-
-const X_MASK: usize = (X_AXIS_SIZE - 1) << X_SHIFT;
-const Z_MASK: usize = (Z_AXIS_SIZE - 1) << Z_SHIFT;
-const Y_MASK: usize = Y_AXIS_SIZE - 1;
-
-#[cfg(feature = "mem_alloc")]
-pub static ALLOC_COUNT: once_cell::sync::Lazy<std::sync::atomic::AtomicUsize> =
-    once_cell::sync::Lazy::new(std::sync::atomic::AtomicUsize::default);
-
-pub const SIDE_COUNT: usize = 4;
-
-#[derive(Clone, Copy, Debug, PartialEq, Eq, PartialOrd, Hash, Default, Serialize, Deserialize)]
-pub enum ChunkSide {
-    #[default]
-    Right = 0,
-    Left = 1,
-    Front = 2,
-    Back = 3,
-}
-
-pub const SIDES: [ChunkSide; SIDE_COUNT] = [
-    ChunkSide::Right,
-    ChunkSide::Left,
-    ChunkSide::Front,
-    ChunkSide::Back,
-];
-
-impl ChunkSide {
-    // pub const fn opposite(&self) -> ChunkSide {
-    //     match self {
-    //         ChunkSide::Right => ChunkSide::Left,
-    //         ChunkSide::Left => ChunkSide::Right,
-    //         ChunkSide::Front => ChunkSide::Back,
-    //         ChunkSide::Back => ChunkSide::Front,
-    //     }
-    // }
-
-    pub const fn index(&self) -> usize {
-        match self {
-            ChunkSide::Right => 0,
-            ChunkSide::Left => 1,
-            ChunkSide::Front => 2,
-            ChunkSide::Back => 3,
-        }
-    }
-
-    pub const fn dir(&self) -> IVec2 {
-        match self {
-            ChunkSide::Right => IVec2::X,
-            ChunkSide::Left => IVec2::NEG_X,
-            ChunkSide::Front => IVec2::Y,
-            ChunkSide::Back => IVec2::NEG_Y,
-        }
-    }
-
-    // pub const fn from_dir(dir: IVec2) -> ChunkSide {
-    //     if dir.x == 1 {
-    //         ChunkSide::Right
-    //     } else if dir.x == -1 {
-    //         ChunkSide::Left
-    //     } else if dir.y == 1 {
-    //         ChunkSide::Front
-    //     } else if dir.y == -1 {
-    //         ChunkSide::Back
-    //     } else {
-    //         panic!("Invalid direction received")
-    //     }
-    // }
-
-    pub const fn from_voxel_side(side: voxel::Side) -> Option<ChunkSide> {
-        match side {
-            voxel::Side::Right => Some(Self::Right),
-            voxel::Side::Left => Some(Self::Left),
-            voxel::Side::Front => Some(Self::Front),
-            voxel::Side::Back => Some(Self::Back),
-            _ => None,
-        }
-    }
-}
-
-pub trait ChunkStorageType:
-    Clone + Copy + core::fmt::Debug + Default + PartialEq + PartialOrd
-{
-}
-
-impl ChunkStorageType for u8 {}
-impl ChunkStorageType for voxel::Kind {}
-impl ChunkStorageType for voxel::Light {}
-impl ChunkStorageType for voxel::FacesOcclusion {}
-impl ChunkStorageType for voxel::FacesSoftLight {}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct ChunkStorage<T>(Vec<T>);
-
-impl<T: ChunkStorageType> Default for ChunkStorage<T> {
-    fn default() -> Self {
-        Self::new(vec![T::default(); BUFFER_SIZE])
-    }
-}
-
-impl<T: ChunkStorageType> PartialEq for ChunkStorage<T> {
-    fn eq(&self, other: &Self) -> bool {
-        self.0 == other.0
-    }
-}
-
-impl<T: ChunkStorageType> ChunkStorage<T> {
-    fn new(buffer: Vec<T>) -> Self {
-        #[cfg(feature = "mem_alloc")]
-        ALLOC_COUNT.fetch_add(1, std::sync::atomic::Ordering::AcqRel);
-
-        Self(buffer)
-    }
-
-    pub fn get(&self, voxel: Voxel) -> T {
-        self.0[to_index(voxel)]
-    }
-
-    pub fn set(&mut self, voxel: Voxel, value: T) {
-        self.0[to_index(voxel)] = value;
-    }
-
-    // pub fn fill(&mut self, value: T) {
-    //     self.0.fill(value);
-    // }
-
-    pub fn iter(&self) -> impl Iterator<Item = &T> {
-        self.0.iter()
-    }
-
-    // pub fn is_default(&self) -> bool {
-    //     // TODO: Add a clever way to check if ChunkStorage wasn't initialized;
-    //     self.is_all(T::default())
-    // }
-
-    // pub fn is_all(&self, value: T) -> bool {
-    //     self.iter().all(|t| *t == value)
-    // }
-
-    // pub fn copy_from(&mut self, other: &Self) {
-    //     self.0.copy_from_slice(&other.0);
-    // }
-}
-
-impl<T: ChunkStorageType> std::ops::Index<usize> for ChunkStorage<T> {
-    type Output = T;
-
-    fn index(&self, index: usize) -> &Self::Output {
-        debug_assert!(index < BUFFER_SIZE);
-        &self.0[index]
-    }
-}
-
-impl<T: ChunkStorageType> std::ops::IndexMut<usize> for ChunkStorage<T> {
-    fn index_mut(&mut self, index: usize) -> &mut Self::Output {
-        debug_assert!(index < BUFFER_SIZE);
-        &mut self.0[index]
-    }
-}
-
-#[cfg(feature = "mem_alloc")]
-impl<T: ChunkStorageType> Drop for ChunkStorage<T> {
-    fn drop(&mut self) {
-        ALLOC_COUNT.fetch_sub(1, std::sync::atomic::Ordering::AcqRel);
-    }
-}
-
-pub trait GetChunkStorage<'a, T: ChunkStorageType + 'a>:
-    Fn(Chunk) -> Option<&'a ChunkStorage<T>> + Copy
-{
-}
-
-impl<'a, T: ChunkStorageType + 'a, F: Copy> GetChunkStorage<'a, T> for F where
-    F: Fn(Chunk) -> Option<&'a ChunkStorage<T>>
-{
-}
-
-pub trait GetChunkStorageMut<'a, T: ChunkStorageType + 'a>:
-    FnMut(Chunk) -> Option<&'a mut ChunkStorage<T>> + Copy
-{
-}
-
-impl<'a, T: ChunkStorageType + 'a, F: Copy> GetChunkStorageMut<'a, T> for F where
-    F: FnMut(Chunk) -> Option<&'a mut ChunkStorage<T>>
-{
-}
-
-impl ChunkStorage<voxel::Light> {
-    pub fn set_type(&mut self, voxel: Voxel, ty: voxel::LightTy, intensity: u8) {
-        let mut light = self.get(voxel);
-        light.set(ty, intensity);
-        self.set(voxel, light);
-    }
-}
-
-#[inline]
-pub fn to_index(voxel: Voxel) -> usize {
-    (voxel.x << X_SHIFT | voxel.y << Y_SHIFT | voxel.z << Z_SHIFT) as usize
-}
-
-#[inline]
-pub fn from_index(index: usize) -> Voxel {
-    Voxel::new(
-        ((index & X_MASK) >> X_SHIFT) as i32,
-        ((index & Y_MASK) >> Y_SHIFT) as i32,
-        ((index & Z_MASK) >> Z_SHIFT) as i32,
-    )
-}
-
-pub fn voxels() -> impl Iterator<Item = Voxel> {
-    (0..BUFFER_SIZE).map(from_index)
-}
-
-pub fn top_voxels() -> impl Iterator<Item = Voxel> {
-    (0..=X_END).flat_map(|x| (0..=Z_END).map(move |z| Voxel::new(x, Y_END, z)))
-}
-
-#[inline]
-pub fn is_inside(voxel: Voxel) -> bool {
-    voxel.x >= 0
-        && voxel.x < X_AXIS_SIZE as i32
-        && voxel.z >= 0
-        && voxel.z < Z_AXIS_SIZE as i32
-        && voxel.y >= 0
-        && voxel.y < Y_AXIS_SIZE as i32
-}
-
-#[cfg(test)]
-pub fn is_at_edge(voxel: Voxel) -> bool {
-    voxel.x == 0
-        || voxel.y == 0
-        || voxel.z == 0
-        || voxel.x == (X_AXIS_SIZE - 1) as i32
-        || voxel.y == (Y_AXIS_SIZE - 1) as i32
-        || voxel.z == (Z_AXIS_SIZE - 1) as i32
-}
-
-pub fn to_world(chunk: Chunk) -> Vec3 {
-    Vec3::new(
-        X_AXIS_SIZE as f32 * chunk.0.x as f32,
-        0.0,
-        Z_AXIS_SIZE as f32 * chunk.0.y as f32,
-    )
-}
-
-pub fn to_chunk(world: Vec3) -> Chunk {
-    Chunk::new(
-        (world.x / X_AXIS_SIZE as f32).floor() as i32,
-        (world.z / Z_AXIS_SIZE as f32).floor() as i32,
-    )
-}
-
-pub fn overlap_voxel(voxel: Voxel) -> (IVec2, Voxel) {
-    debug_assert!(!is_inside(voxel), "Voxel {voxel} does't overlap");
-    debug_assert!(
-        voxel.y >= 0 && voxel.y < Y_AXIS_SIZE as i32,
-        "Can't overlap up or down. There is never chunk above or bellow"
-    );
-
-    let overlapping_voxel = math::euclid_rem(
-        voxel,
-        IVec3::new(X_AXIS_SIZE as i32, Y_AXIS_SIZE as i32, Z_AXIS_SIZE as i32),
-    );
-
-    let overlapping_dir = (
-        if voxel.x < 0 {
-            -1
-        } else if voxel.x >= X_AXIS_SIZE as i32 {
-            1
-        } else {
-            0
-        },
-        if voxel.z < 0 {
-            -1
-        } else if voxel.z >= Z_AXIS_SIZE as i32 {
-            1
-        } else {
-            0
-        },
-    )
-        .into();
-
-    (overlapping_dir, overlapping_voxel)
-}
-
-#[cfg(test)]
-mod tests {
-    use bevy::math::IVec3;
-    use rand::Rng;
-
-    use super::*;
-
-    #[test]
-    fn from_index() {
-        assert_eq!(Voxel::new(0, 0, 0), super::from_index(0));
-        assert_eq!(Voxel::new(0, 1, 0), super::from_index(1));
-        assert_eq!(Voxel::new(0, 2, 0), super::from_index(2));
-
-        assert_eq!(
-            Voxel::new(0, 0, 1),
-            super::from_index(super::Y_AXIS_SIZE),
-            "X >> Z >> Y, so one Z unit should be a full Y axis"
-        );
-        assert_eq!(
-            Voxel::new(0, 1, 1),
-            super::from_index(super::Y_AXIS_SIZE + 1)
-        );
-        assert_eq!(
-            Voxel::new(0, 2, 1),
-            super::from_index(super::Y_AXIS_SIZE + 2)
-        );
-
-        assert_eq!(
-            Voxel::new(1, 0, 0),
-            super::from_index(super::Y_AXIS_SIZE * super::Z_AXIS_SIZE)
-        );
-        assert_eq!(
-            Voxel::new(1, 1, 0),
-            super::from_index(super::Y_AXIS_SIZE * super::Z_AXIS_SIZE + 1)
-        );
-        assert_eq!(
-            Voxel::new(1, 2, 0),
-            super::from_index(super::Y_AXIS_SIZE * super::Z_AXIS_SIZE + 2)
-        );
-
-        assert_eq!(
-            Voxel::new(1, 0, 1),
-            super::from_index(super::Y_AXIS_SIZE * super::Z_AXIS_SIZE + super::Y_AXIS_SIZE)
-        );
-        assert_eq!(
-            Voxel::new(1, 1, 1),
-            super::from_index(super::Y_AXIS_SIZE * super::Z_AXIS_SIZE + super::Y_AXIS_SIZE + 1)
-        );
-        assert_eq!(
-            Voxel::new(1, 2, 1),
-            super::from_index(super::Y_AXIS_SIZE * super::Z_AXIS_SIZE + super::Y_AXIS_SIZE + 2)
-        );
-    }
-
-    #[test]
-    fn to_index() {
-        assert_eq!(super::to_index((0, 0, 0).into()), 0);
-        assert_eq!(super::to_index((0, 1, 0).into()), 1);
-        assert_eq!(super::to_index((0, 2, 0).into()), 2);
-
-        assert_eq!(super::to_index((0, 0, 1).into()), super::Y_AXIS_SIZE);
-        assert_eq!(super::to_index((0, 1, 1).into()), super::Y_AXIS_SIZE + 1);
-        assert_eq!(super::to_index((0, 2, 1).into()), super::Y_AXIS_SIZE + 2);
-
-        assert_eq!(
-            super::to_index((1, 0, 0).into()),
-            super::Y_AXIS_SIZE * super::Z_AXIS_SIZE
-        );
-        assert_eq!(
-            super::to_index((1, 1, 0).into()),
-            super::Y_AXIS_SIZE * super::Z_AXIS_SIZE + 1
-        );
-        assert_eq!(
-            super::to_index((1, 2, 0).into()),
-            super::Y_AXIS_SIZE * super::Z_AXIS_SIZE + 2
-        );
-
-        assert_eq!(
-            super::to_index((1, 0, 1).into()),
-            super::Y_AXIS_SIZE * super::Z_AXIS_SIZE + super::Y_AXIS_SIZE
-        );
-        assert_eq!(
-            super::to_index((1, 1, 1).into()),
-            super::Y_AXIS_SIZE * super::Z_AXIS_SIZE + super::Y_AXIS_SIZE + 1
-        );
-        assert_eq!(
-            super::to_index((1, 2, 1).into()),
-            super::Y_AXIS_SIZE * super::Z_AXIS_SIZE + super::Y_AXIS_SIZE + 2
-        );
-    }
-
-    #[test]
-    fn to_world() {
-        use super::*;
-
-        assert_eq!(
-            super::to_world(Chunk::new(0, -2)),
-            Vec3::new(0.0, 0.0, super::Z_AXIS_SIZE as f32 * -2.0)
-        );
-        assert_eq!(
-            super::to_world(Chunk::new(3, 1)),
-            Vec3::new(
-                super::X_AXIS_SIZE as f32 * 3.0,
-                0.0,
-                super::Z_AXIS_SIZE as f32 * 1.0
-            )
-        );
-    }
-
-    // #[test]
-    // fn to_chunk() {
-    //     use super::*;
-    //
-    //     assert_eq!(
-    //         Chunk::new(0, -2),
-    //         super::to_chunk(Vec3::new(0.0, 0.0, -2.0))
-    //     );
-    //     assert_eq!(Chunk::new(0, 0), super::to_chunk(Vec3::new(0.0, 0.0, 0.0)));
-    // }
-
-    #[test]
-    fn voxels() {
-        let mut first = None;
-        let mut last = IVec3::ZERO;
-
-        for pos in super::voxels() {
-            assert!(pos.x >= 0 && pos.x < super::X_AXIS_SIZE as i32);
-            assert!(pos.y >= 0 && pos.y < super::Y_AXIS_SIZE as i32);
-            assert!(pos.z >= 0 && pos.z < super::Z_AXIS_SIZE as i32);
-
-            if first.is_none() {
-                first = Some(pos);
-            }
-            last = pos;
-        }
-
-        assert_eq!(first, Some(IVec3::ZERO));
-        assert_eq!(
-            last,
-            (
-                X_AXIS_SIZE as i32 - 1,
-                Y_AXIS_SIZE as i32 - 1,
-                Z_AXIS_SIZE as i32 - 1
-            )
-                .into()
-        );
-    }
-
-    #[test]
-    fn top_voxels() {
-        let top_voxels = super::top_voxels().collect::<Vec<_>>();
-
-        assert_eq!(top_voxels.len(), X_AXIS_SIZE * Z_AXIS_SIZE);
-        top_voxels
-            .into_iter()
-            .for_each(|voxel| assert_eq!(voxel.y, Y_END));
-    }
-
-    #[test]
-    fn set_get() {
-        let mut chunk = ChunkStorage::<u8>::default();
-
-        let mut rnd = rand::thread_rng();
-        for v in super::voxels() {
-            let k = rnd.gen::<u8>();
-            chunk.set(v, k);
-            assert_eq!(k, chunk.get(v));
-        }
-    }
-
-    // #[test]
-    // fn is_default() {
-    //     impl ChunkStorageType for [u8; 3] {}
-    //
-    //     let mut chunk = ChunkStorage::<[u8; 3]>::default();
-    //     assert!(chunk.is_default());
-    //
-    //     chunk.set((1, 1, 1).into(), [1; 3]);
-    //
-    //     assert!(!chunk.is_default());
-    // }
-
-    #[test]
-    fn is_at_edge() {
-        let voxel = (1, 1, 1).into();
-        assert!(!super::is_at_edge(voxel));
-
-        let voxel = (1, 0, 1).into();
-        assert!(super::is_at_edge(voxel));
-
-        let voxel = (1, Y_END, 1).into();
-        assert!(super::is_at_edge(voxel));
-
-        let voxel = (0, 0, 0).into();
-        assert!(super::is_at_edge(voxel));
-
-        let voxel = (2, 1, 14).into();
-        assert!(!super::is_at_edge(voxel));
-    }
-}
