+++ crates/world_server/src/genesis.rs
@@ -1,9 +1,7 @@
 use bracket_noise::prelude::*;
-use projekto_core::voxel;
-
-use crate::{
-    chunk::{self, ChunkStorage},
-    Chunk,
+use projekto_core::{
+    chunk::{self, Chunk, ChunkStorage},
+    voxel,
 };
 
 /// Generates a new chunk filling it with [`ChunkKind`] randomly generated by seeded noise
